requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.6.0
anthropic==0.25.1
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
filelock==3.13.4
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Migrate==4.0.5
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.1.16
langchain-anthropic==0.1.8
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langsmith==0.1.47
Mako==1.3.3
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==1.26.4
openai==1.13.3
orjson==3.10.0
packaging==23.2
pbr==6.0.0
platformdirs==4.2.0
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
stevedore==5.2.0
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.15.2
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
urllib3==2.2.1
virtualenv==20.25.1
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
Werkzeug==3.0.2
yarl==1.9.4
